moving package metrics to server
