+++ server/metrics/enabled_test.go
